Move TaskStreamPlugin to diagnostics
